# Conflicts:
#	Book/target/classes/application.yml
#	Book/target/classes/onlinebookstore/AbstractEvent$1.class
#	Book/target/classes/onlinebookstore/AbstractEvent.class
#	Book/target/classes/onlinebookstore/Book.class
#	Book/target/classes/onlinebookstore/BookApplication.class
#	Book/target/classes/onlinebookstore/BookController.class
#	Book/target/classes/onlinebookstore/BookRegistred.class
#	Book/target/classes/onlinebookstore/BookRepository.class
#	Book/target/classes/onlinebookstore/OrderCancelled.class
#	Book/target/classes/onlinebookstore/PolicyHandler.class
#	Book/target/classes/onlinebookstore/StockDecreased.class
#	Book/target/classes/onlinebookstore/StockIncreased.class

+++ customer/src/main/java/onlinebookstore/Ordered.java
@@ -3,7 +3,7 @@ package onlinebookstore;
 public class Ordered extends AbstractEvent {
 
     private Long orderId;
-    private Integer bookId;
+    private Long bookId;
     private Integer qty;
     private Integer price;
     private Integer paymentId;
@@ -17,11 +17,11 @@ public class Ordered extends AbstractEvent {
     public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
-    public Integer getBookId() {
+    public Long getBookId() {
         return bookId;
     }
 
-    public void setBookId(Integer bookId) {
+    public void setBookId(Long bookId) {
         this.bookId = bookId;
     }
     public Integer getQty() {
